move CoordinateGrids to new module. relocate geolocation.py
